+++ src/models/RootStateTypes.tsx
@@ -1,10 +1,10 @@
 import { Provider as Provider, Client as Client } from "./UsersModels";
+import { UserState } from "./UsersModels";
 
 interface RootState {
-  user: any;
+  user: UserState;
   clients: { clients: Client[] };
   providers: { providers: Provider[] };
 }
-
 
 export default RootState;
