+++ src/models/UsersModels.tsx
@@ -35,3 +35,35 @@ interface Client {
 }
 
 export type { Client };
+
+interface UserState {
+  user: {
+    username: string;
+    userType: string;
+
+    name: string;
+    phone: string;
+    email: string;
+    photo: string;
+
+    address?: string;
+
+    company?: {
+      name: string;
+      logo: string;
+      phone: string;
+      email: string;
+    };
+    service?: {
+      type: string;
+    };
+    responsibleCompany?: {
+      name: string;
+      phone: string;
+      email: string;
+      photo: string;
+    };
+  } | null;
+}
+
+export type { UserState };
